+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Added support for `XChainModifyBridge` flag maps
+- Fixed `XChainModifyBridge` validation to allow just clearing of `MinAccountCreateAmount`
 
 ## [2.5.0] - 2023-11-30