+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [[Unreleased]]
+- Included `ctid` field in the `tx` request.
+
+### Fixed
+- Added support for `XChainModifyBridge` flag maps (fixing an issue with `NFTokenCreateOffer` flag names)
+- Fixed `XChainModifyBridge` validation to allow just clearing of `MinAccountCreateAmount`
 
 ## [2.5.0] - 2023-11-30