+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for `XChainModifyBridge` flag maps (fixing an issue with `NFTokenCreateOffer` flag names)
 - Fixed `XChainModifyBridge` validation to allow just clearing of `MinAccountCreateAmount`
 - Added support for IDE auto-completion of Ledger Object constructors using KW_ONLY data_class attribute
+- Currency codes with special characters not being allowed by IssuedCurrency objects.
 
 ## [2.5.0] - 2023-11-30