+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [[Unreleased]]
 - Add support for the DeliverMax field in Payment transactions
+- Included `ctid` field in the `tx` request.
+
+### Fixed
+- Added support for `XChainModifyBridge` flag maps (fixing an issue with `NFTokenCreateOffer` flag names)
+- Fixed `XChainModifyBridge` validation to allow just clearing of `MinAccountCreateAmount`
+- Currency codes with special characters not being allowed by IssuedCurrency objects.
 
 ## [2.5.0] - 2023-11-30