+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Included `ctid` field in the `tx` request.
 
 ### Added
-- Support for the PriceOracle amendment (XLS-47d).
+- Support for the Price Oracles amendment (XLS-47).
 
 ### Fixed
 - Added support for `XChainModifyBridge` flag maps (fixing an issue with `NFTokenCreateOffer` flag names)